replace relationship_count column in Roles CRUD with a text one
Fixes https://github.com/Laravel-Backpack/CRUD/issues/3227 . Because we were using the `relationship_count` column, for every Role all Users were loaded onto the page, in order to count them. With a big Users table this quickly got out of hand, causing performance issues (15 sec load time for example). This PR replaces the `relationship_count` column with a `text` column, and calculates the `users_count` ahead of time, using a method Laravel provides that is A LOT faster and does not load the Users at all.